cooperativa_backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
sqlparse==0.4.4
pytz==2023.3
requests==2.31.0